fix unmount devices when scrpit chroot
When the chroot destination is a folder and this folder is already a
mounted partition (ie: a loop mount of an fs file to create a rootfs),
if the install script is run chrooted (--script-chroot), than at the end
of the execution of the install script we try to unmount recursively
everything that is mounted in the destination folder AND the
destination, which will cause the end of this script to fail.

The solution I am proposing is to create another bash function to
unmount only required devices for the chroot (/proc, /dev and /sys).