Install ALPINE_BASE_PKGS first if requested
When a package has a pre-install script that adds users, the pre-install
actually requires alpine-baselayout-data/busybox to successfully work,
but this dependency isn't written anywhere.

After bringing this up on IRC rather than adding the dependency everywhere
it was decided to make sure alpine-baselayout is installed first instead,
so do this.

--initdb can be used multiple times with no ill side effects, but
--updatedb is annoying and not actually useful (will automatically
refresh every 4 hours, but starting from an empty directory it'll
always fetch it anyway), so just remove it.

Co-Authored-By: Jakub Jirutka <jakub@jirutka.cz>